--- src/main/java/org/example/mariajeu/domain/AtmosphereStatus.java
@@ -1,7 +1,0 @@
-package org.example.mariajeu.domain;
-
-public enum AtmosphereStatus {
-    DATE,
-    ANNIVERSARY,
-    BLINDDATE
-}
