--- src/main/java/org/example/mariajeu/domain/ReservationStatus.java
@@ -1,7 +1,0 @@
-package org.example.mariajeu.domain;
-
-
-public enum ReservationStatus {
-    AVAILABLE,   // 예약 가능
-    COMPLETED // 예약 완료
-}
