+++ src/main/java/org/example/mariajeu/domain/restaurantDomain/AtmosphereStatus.java
@@ -1,0 +1,7 @@
+package org.example.mariajeu.domain.restaurantDomain;
+
+public enum AtmosphereStatus {
+    DATE,
+    ANNIVERSARY,
+    BLINDDATE
+}
