+++ src/main/java/org/example/mariajeu/domain/restaurantDomain/ReservationStatus.java
@@ -1,0 +1,7 @@
+package org.example.mariajeu.domain.restaurantDomain;
+
+
+public enum ReservationStatus {
+    AVAILABLE,   // 예약 가능
+    COMPLETED // 예약 완료
+}
