--- src/main/java/org/example/mariajeu/service/RestaurantNotFoundException.java
@@ -1,6 +1,0 @@
-package org.example.mariajeu.service;
-
-public class RestaurantNotFoundException extends Throwable {
-    public RestaurantNotFoundException(String s) {
-    }
-}
