+++ src/main/java/org/example/mariajeu/service/restaurantService/RestaurantNotFoundException.java
@@ -1,0 +1,6 @@
+package org.example.mariajeu.service.restaurantService;
+
+public class RestaurantNotFoundException extends Throwable {
+    public RestaurantNotFoundException(String s) {
+    }
+}
