+++ src/test/java/org/example/mariajeu/repository/RestaurantRepositoryTest.java
@@ -1,13 +1,8 @@
 package org.example.mariajeu.repository;
 
-import org.example.mariajeu.domain.Restaurant;
-import org.junit.jupiter.api.Test;
+import org.example.mariajeu.repository.restaurantRepository.RestaurantRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-
-import java.util.List;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 @DataJpaTest
 public class RestaurantRepositoryTest {
